Rename TodolistMenu
